ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pika==1.3.2
boto3==1.35.24
pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
pydantic==2.9.2
pydantic-settings==2.5.2
